+++ plugins/drivers/server.go
@@ -43,6 +43,7 @@ func (b *driverPluginServer) Capabilities(ctx context.Context, req *proto.Capabi
 			Exec:                  caps.Exec,
 			MustCreateNetwork:     caps.MustInitiateNetwork,
 			NetworkIsolationModes: []proto.NetworkIsolationSpec_NetworkIsolationMode{},
+			RemoteTasks:           caps.RemoteTasks,
 		},
 	}